# docs/swagger.yaml
basePath: /v1
definitions:
  api.ErrorDetail:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      success:
        type: boolean
    type: object
...
  api.ResponseMetadata:
    properties:
      timestamp:
        type: string
      trace_id:
        type: string
    type: object
  api.SuccessResponse:
    properties:
      data: {}
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      success:
        type: boolean
    type: object
...
  services.CreateSubAccountRequest:
    properties:
      name:
        example: BTC Trading Account
        maxLength: 100
        minLength: 1
        type: string
      symbol:
        example: BTC/USDT
        maxLength: 20
        minLength: 1
        type: string
      trading_id:
        example: 453f0347-3959-49de-8e3f-1cf7c8e0827c
        type: string
    required:
    - name
    - symbol
    - trading_id
    type: object
...
  services.SubAccountResponse:
    properties:
      balance:
        type: number
      created_at:
        type: string
      id:
        type: string
      info:
        additionalProperties: true
        type: object
      name:
        type: string
      symbol:
        type: string
      trading_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  services.UpdateBalanceRequest:
    properties:
      amount:
        example: 500.25
        type: number
      direction:
        enum:
        - credit
        - debit
        example: credit
        type: string
      info:
        additionalProperties: true
        type: object
      reason:
        example: Successful BTC/USDT trade profit
        maxLength: 255
        minLength: 1
        type: string
    required:
    - amount
    - direction
    - reason
    type: object
  services.UpdateSubAccountRequest:
    properties:
      balance:
        example: 1250.75
        minimum: 0
        type: number
      name:
        example: ETH Trading Account
        maxLength: 100
        minLength: 1
        type: string
      symbol:
        example: ETH/USD
        maxLength: 20
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@tiris.ai
    name: Tiris API Support
    url: https://tiris.ai/support
  description: A Go-based microservice for quantitative trading data management, providing
    RESTful APIs for user management, trading integration, and trading operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://tiris.ai/terms
  title: Tiris Backend API
  version: "1.0"
paths:
...
  /sub-accounts:
    get:
      description: Retrieves all sub-account configurations for the authenticated
        user, optionally filtered by trading
      parameters:
      - description: Filter by trading ID
        in: query
        name: trading_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.SubAccountResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user sub-accounts
      tags:
      - SubAccounts
    post:
      consumes:
      - application/json
      description: Creates a new sub-account for the authenticated user
      parameters:
      - description: Create sub-account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateSubAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.SubAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new sub-account
      tags:
      - SubAccounts
  /sub-accounts/{id}:
    delete:
      description: Deletes a sub-account configuration (must belong to authenticated
        user)
      parameters:
      - description: Sub-account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete sub-account
      tags:
      - SubAccounts
    get:
      description: Retrieves a specific sub-account configuration by ID (must belong
        to authenticated user)
      parameters:
      - description: Sub-account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SubAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sub-account by ID
      tags:
      - SubAccounts
    put:
      consumes:
      - application/json
      description: Updates an existing sub-account configuration (must belong to authenticated
        user)
      parameters:
      - description: Sub-account ID
        in: path
        name: id
        required: true
        type: string
      - description: Update sub-account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateSubAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SubAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update sub-account
      tags:
      - SubAccounts
  /sub-accounts/{id}/balance:
    put:
      consumes:
      - application/json
      description: Updates sub-account balance with proper logging (must belong to
        authenticated user)
      parameters:
      - description: Sub-account ID
        in: path
        name: id
        required: true
        type: string
      - description: Update balance request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SubAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update sub-account balance
      tags:
      - SubAccounts
  /sub-accounts/symbol/{symbol}:
    get:
      description: Retrieves all sub-accounts for a specific trading symbol (must
        belong to authenticated user)
      parameters:
      - description: Trading symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.SubAccountResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sub-accounts by symbol
      tags:
      - SubAccounts
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT Bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...